Create 2019-07-18-BPS - ds-recovery checkpoint.ipynb
This notebook works so i am checking it in as its own version.  It got 100% validation accuracy, 83% test accuracy, though at this stage validation and test are functionally identical so it doesn't matter too much which is which.

Next step is to make and test bootstrapping configuations to get confidence interval estimates, as well as implement kfold.